feat(apps/frontend-manage): upgrade live quiz creation editor with new components (#4151)
