+++ .github/workflows/cypress-testing.yml
@@ -3,7 +3,6 @@ on:
   push:
     branches: ["v3", "v3*"]
   pull_request:
-    branches: ["v3", "v3*"]
     types: [opened, synchronize, reopened]
   # workflow_dispatch:
