+++ apps/backend-docker/package.json
@@ -52,7 +52,7 @@
     "@types/passport-jwt": "^3.0.6",
     "@types/ramda": "^0.29.3",
     "@types/ws": "^8.5.9",
-    "@uzh-bf/design-system": "2.11.0",
+    "@uzh-bf/design-system": "2.13.1",
     "axios": "1.4.0",
     "cross-env": "7.0.3",
     "eslint": "8.45.0",
