+++ apps/frontend-manage/src/components/sessions/cockpit/SessionBlock.tsx
@@ -99,7 +99,7 @@ function SessionBlock({
         block.numOfParticipants !== null ? (
           <div className="flex flex-row items-center">
             <span className="font-bold">
-              {t('manage.cockpit.blockN', {
+              {t('shared.generic.blockN', {
                 number: block.order! + 1,
               })}
             </span>
@@ -107,7 +107,7 @@ function SessionBlock({
             <FontAwesomeIcon icon={faUserGroup} className="w-4 ml-1" />
           </div>
         ) : (
-          <div>{t('manage.cockpit.blockN', { number: block.order! + 1 })}</div>
+          <div>{t('shared.generic.blockN', { number: block.order! + 1 })}</div>
         )}
 
         {block.timeLimit && (
