+++ apps/frontend-manage/src/components/sessions/creation/MultistepWizard.tsx
@@ -4,7 +4,11 @@ import {
   faSave,
 } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { ElementOrderType, ParameterType } from '@klicker-uzh/graphql/dist/ops'
+import {
+  ElementOrderType,
+  ElementType,
+  ParameterType,
+} from '@klicker-uzh/graphql/dist/ops'
 import { Button, H2, Workflow } from '@uzh-bf/design-system'
 import { Form, Formik } from 'formik'
 import { useTranslations } from 'next-intl'
@@ -54,13 +58,22 @@ interface CommonFormValues {
   multiplier: string
 }
 
+export interface LiveQuizBlockFormValues {
+  questionIds: number[]
+  titles: string[]
+  types: ElementType[]
+  timeLimit: number
+}
+
+export interface LiveQuizBlockErrorValues {
+  questionIds: string[]
+  titles: string[]
+  types: string[]
+  timeLimit: string
+}
+
 export interface LiveSessionFormValues extends CommonFormValues {
-  blocks: {
-    questionIds: number[]
-    titles: string[]
-    types: []
-    timeLimit: number
-  }[]
+  blocks: LiveQuizBlockFormValues[]
   isGamificationEnabled: boolean
   isConfusionFeedbackEnabled: boolean
   isLiveQAEnabled: boolean
@@ -142,7 +155,8 @@ function MultistepWizard({
     values:
       | LiveSessionFormValues
       | MicroLearningFormValues
-      | PracticeQuizFormValues,
+      | PracticeQuizFormValues
+      | GroupActivityFormValues,
     bag: any
   ) => {
     if (step.props.onSubmit) {
@@ -165,30 +179,65 @@ function MultistepWizard({
       isInitialValid={initialValid}
     >
       {({ values, isSubmitting, isValid, resetForm, validateForm }) => (
-        <Form className="h-full overflow-y-auto">
+        <Form className="h-full w-full flex flex-col">
           <Validator stepNumber={stepNumber} validateForm={validateForm} />
-          <div className="flex flex-row items-end gap-8">
+          <div className="flex flex-row items-end gap-8 h-6">
             <H2 className={{ root: 'flex flex-none m-0 items-end' }}>
               {editMode
                 ? t('manage.questionForms.editElement', { element: title })
                 : t('manage.questionForms.createElement', { element: title })}
             </H2>
             <Workflow
+              minimal
+              showTooltipSymbols
               items={workflowItems}
               onClick={(_, ix) => setStepNumber(ix)}
               activeIx={stepNumber}
               // TODO: choose optimal disabled logic - allow to jump between 3 and 1 if all valid
               disabledFrom={isValid ? stepNumber + 2 : stepNumber + 1}
-              minimal
-              showTooltipSymbols
               className={{
-                item: 'last:rounded-r-md first:rounded-l-md',
+                item: 'last:rounded-r-md first:rounded-l-md hidden md:flex',
               }}
             />
           </div>
 
-          <div className="flex flex-col justify-between gap-1 py-4">
-            {!isCompleted && <>{step}</>}
+          <div className="flex flex-1 w-full justify-between gap-1 py-4">
+            {!isCompleted && (
+              <div className="flex flex-col justify-between w-full h-full">
+                <div>{step}</div>
+                <div className="flex flex-row justify-between pt-2">
+                  <Button
+                    className={{ root: 'border-red-400' }}
+                    onClick={() => onCloseWizard()}
+                    data={{ cy: 'cancel-session-creation' }}
+                  >
+                    <FontAwesomeIcon icon={faCancel} />
+                    <div>
+                      {editMode
+                        ? t('manage.questionForms.cancelEditing')
+                        : t('manage.questionForms.cancelCreation')}
+                    </div>
+                  </Button>
+                  <Button
+                    disabled={isSubmitting || !isValid}
+                    type="submit"
+                    data={{ cy: 'next-or-submit' }}
+                    className={{ root: 'w-max self-end' }}
+                  >
+                    {stepNumber === steps.length - 1 ? (
+                      <FontAwesomeIcon icon={faSave} />
+                    ) : (
+                      <FontAwesomeIcon icon={faArrowRight} />
+                    )}
+                    {stepNumber === steps.length - 1
+                      ? editMode
+                        ? t('shared.generic.save')
+                        : t('shared.generic.create')
+                      : t('shared.generic.continue')}
+                  </Button>
+                </div>
+              </div>
+            )}
 
             {isCompleted && (
               <CompletionStep
@@ -202,40 +251,6 @@ function MultistepWizard({
               >
                 {customCompletionAction}
               </CompletionStep>
-            )}
-
-            {!isCompleted && (
-              <div className="flex flex-row justify-between pt-2">
-                <Button
-                  className={{ root: 'border-red-400' }}
-                  onClick={() => onCloseWizard()}
-                  data={{ cy: 'cancel-session-creation' }}
-                >
-                  <FontAwesomeIcon icon={faCancel} />
-                  <div>
-                    {editMode
-                      ? t('manage.questionForms.cancelEditing')
-                      : t('manage.questionForms.cancelCreation')}
-                  </div>
-                </Button>
-                <Button
-                  disabled={isSubmitting || !isValid}
-                  type="submit"
-                  data={{ cy: 'next-or-submit' }}
-                  className={{ root: 'w-max self-end' }}
-                >
-                  {stepNumber === steps.length - 1 ? (
-                    <FontAwesomeIcon icon={faSave} />
-                  ) : (
-                    <FontAwesomeIcon icon={faArrowRight} />
-                  )}
-                  {stepNumber === steps.length - 1
-                    ? editMode
-                      ? t('shared.generic.save')
-                      : t('shared.generic.create')
-                    : t('shared.generic.continue')}
-                </Button>
-              </div>
             )}
           </div>
         </Form>
