+++ apps/frontend-manage/src/components/sessions/creation/PropertyList.tsx
@@ -1,0 +1,37 @@
+import { IconDefinition } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import React from 'react'
+import { twMerge } from 'tailwind-merge'
+
+interface PropertyListProps {
+  elements: {
+    key: string
+    icon: IconDefinition
+    iconColor: string
+    richText: string | React.ReactNode
+  }[]
+}
+
+function PropertyList({ elements }: PropertyListProps) {
+  return (
+    <>
+      {elements.map((element) => (
+        <div
+          className="flex flex-row items-center text-[0.95rem]"
+          key={element.key}
+        >
+          <div className="w-fit">
+            <FontAwesomeIcon
+              icon={element.icon}
+              className={twMerge('w-10 -ml-1.5', element.iconColor)}
+              size="lg"
+            />
+          </div>
+          <div className="leading-5">{element.richText}</div>
+        </div>
+      ))}
+    </>
+  )
+}
+
+export default PropertyList
