+++ apps/office-addin/package.json
@@ -3,7 +3,7 @@
   "version": "3.1.2-rc.2",
   "license": "AGPL-3.0",
   "dependencies": {
-    "@uzh-bf/design-system": "2.11.0",
+    "@uzh-bf/design-system": "2.13.1",
     "core-js": "3.30.2",
     "es6-promise": "4.2.8",
     "formik": "2.4.3",
