+++ packages/lti/package.json
@@ -24,7 +24,7 @@
     "typescript": "5.0.4"
   },
   "peerDependencies": {
-    "@uzh-bf/design-system": "2.11.0",
+    "@uzh-bf/design-system": "2.13.1",
     "next": "14.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0"
