+++ packages/shared-components/package.json
@@ -32,7 +32,7 @@
     "@fortawesome/free-regular-svg-icons": "6.4.2",
     "@fortawesome/free-solid-svg-icons": "6.4.2",
     "@fortawesome/react-fontawesome": "0.2.0",
-    "@uzh-bf/design-system": "2.11.0",
+    "@uzh-bf/design-system": "2.13.1",
     "formik": "2.4.3",
     "js-cookie": "3.0.5",
     "localforage": "1.10.0",
